heredoc 내용확장 완성 및 파일 분리
