+++ Makefile
@@ -6,17 +6,17 @@
 #    By: han-yeseul <han-yeseul@student.42.fr>      +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2022/07/15 21:26:09 by minkyeki          #+#    #+#              #
-#    Updated: 2022/07/28 13:24:47 by han-yeseul       ###   ########.fr        #
+#    Updated: 2022/07/28 16:11:26 by han-yeseul       ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
 NAME			= minishell
 CC				= cc
 
-				#-g3d 옵션 붙이면 오류가 뜬 파일 위치와 줄번호까지 뜸
-CFLAGS		= -Wall -Wextra -Werror -g#3 -fsanitize=address
+# 				#-g3d 옵션 붙이면 오류가 뜬 파일 위치와 줄번호까지 뜸
+# CFLAGS		= -Wall -Wextra -Werror -g3 -fsanitize=address
 
-# CFLAGS			= -Wall -Wextra -Werror
+CFLAGS			= -Wall -Wextra -Werror
 INCLUDE			= include
 
 # GNU readline Flag for MAC
@@ -37,14 +37,15 @@ ITERATOR_DIR	= $(SRC_DIR)/iterator/
 
 # NOTE : Add Source files here
 # ------------------------------------------------------ #
-MAIN_SRC		= minishell prompt helper signal_handle heredoc
+MAIN_SRC		= minishell prompt helper signal_handle heredoc heredoc_utils
 
 LEXER_SRC		= token_create token_modify \
 				  scanner_main scanner_create scanner_function_ptr \
 				  scanner_helper scanner_get_token_1 scanner_get_token_2
 
 PARSER_SRC		= parse_tree parse_tree_utils
-EXECUTER_SRC	= executer token_expand get_full_path redirection subshell
+EXECUTER_SRC	= executer token_expand get_full_path subshell \
+				  redirection redirection_utils exec_command exec_command_utils
 
 BUILTIN_SRC		= environ environ_utils environ_helper cd echo pwd exit
 STRING_SRC		= string_create string_modify string_utils
