+++ src/executer/executer.h
@@ -6,7 +6,7 @@
 /*   By: han-yeseul <han-yeseul@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/17 22:19:07 by minkyeki          #+#    #+#             */
-/*   Updated: 2022/07/27 20:49:29 by minkyeki         ###   ########.fr       */
+/*   Updated: 2022/07/28 14:24:25 by han-yeseul       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,21 +26,21 @@
 # define CMD_STOP_SHELL		(-1)
 
 /**-
- * @ Word Expansion function 
+ * @ Word Expansion function
  * - src/executer/token_expand.c
  **/
 int		expand_tokens(t_list *tokens, t_shell_config *config);
 
-/**- 
+/**-
  * @ Get program's full path.
  * - Returns NULL if program doesn't exist.
  * - src/executer/get_full_path.c
  **/
 char	*get_full_path(char *name, char **envp);
 
-/**- 
+/**-
  * @ Convert token to argv (char **argv)
- * - src/executer/token_expand.c 
+ * - src/executer/token_expand.c
  **/
 char	**get_cmd_argv(t_list *token);
 
@@ -52,7 +52,7 @@ char	**get_cmd_argv(t_list *token);
 /* get_full_path */
 int		is_valid_path(char *path, char *name);
 
-int 	open_redirection(int *pipe_fd, t_list *redir_list);
+int 	open_redirection(int *pipe_fd, t_list *redir_list, t_shell_config *config);
 void	delete_tree_node(t_tree *node, int *status, t_shell_config *config);
 void	execute_node(t_tree *node, int *status, t_shell_config *config);
 void	inorder_recur(t_tree *node, int *status, void (*f)(t_tree *node, int *status, t_shell_config *config), t_shell_config *config);
