+++ src/main/heredoc.c
@@ -1,23 +1,10 @@
 #include <readline/readline.h>
 #include <fcntl.h>
 #include "minishell.h"
+#include "heredoc.h"
 #include "../lexer/token.h"
 
 # define CHILD (0)
-
-static void	expand_quote(t_string *str, t_iterator *iter, char quote_type)
-{
-	char	c;
-
-	while (iter->f_has_next(iter))
-	{
-		c = iter->f_next(iter);
-		if (c == quote_type)
-			return ;
-		else
-			str->f_push_back(str, c);
-	}
-}
 
 void	expand_token(t_token *tok)
 {
@@ -40,32 +27,6 @@ void	expand_token(t_token *tok)
 	}
 	delete_string(&tok->str);
 	tok->str = expanded_str;
-}
-
-/*********************************************/
-
-static bool	is_limiter(const char *line, const char *limiter)
-{
-	if (ft_strlen(line) == ft_strlen(limiter)
-		&& !ft_strncmp(line, limiter, ft_strlen(limiter)))
-		return (true);
-	else
-		return (false);
-}
-
-static char	*readline_prompt_heredoc(void)
-{
-	t_string	*prompt;
-	static char	*line;
-
-	prompt = new_string(64);
-	prompt->f_append(prompt, "\033[31m");
-	prompt->f_append(prompt, "heredoc");
-	prompt->f_append(prompt, "\033[0m");
-	prompt->f_append(prompt, "$ ");
-	line = readline(prompt->text);
-	delete_string(&prompt);
-	return (line);
 }
 
 void	open_heredoc(t_token *tok)
@@ -98,8 +59,6 @@ void	open_heredoc(t_token *tok)
 	}
 }
 
-/*************************************************/
-
 void	set_heredoc(t_list *tokens)
 {
 	t_list	*cur;
@@ -119,5 +78,3 @@ void	set_heredoc(t_list *tokens)
 		cur = cur->next;
 	}
 }
-
-/***************************************/
